Move Sequence implementation into data structures
